Allow skslc to compile one file at a time.
Set `batchCompile = False` to compile each input file individually. This
is slower, but can be useful when attempting to pinpoint which input
file is causing a crash in skslc.

+++ gn/compile_sksl_tests.py
@@ -14,6 +14,22 @@ skslc = sys.argv[1]
 lang = sys.argv[2]
 settings = sys.argv[3]
 inputs = sys.argv[4:]
+batchCompile = True
+
+def executeWorklist(input, worklist):
+    # Invoke skslc, passing in the worklist.
+    worklist.close()
+    try:
+        output = subprocess.check_output([skslc, worklist.name], stderr=subprocess.STDOUT)
+    except subprocess.CalledProcessError as err:
+        if err.returncode != 1:
+            print("### " + input + " skslc error:\n")
+            print("\n".join(err.output.splitlines()))
+            sys.exit(err.returncode)
+        pass  # Compile errors (exit code 1) are expected and normal in test code
+
+    # Delete the worklist file now that execution is complete.
+    os.remove(worklist.name)
 
 def makeEmptyFile(path):
     try:
@@ -70,18 +86,17 @@ for input in inputs:
     else:
         sys.exit("### Expected one of: --fp --glsl --metal --spirv --skvm, got " + lang)
 
-# Invoke skslc, passing in the worklist.
-worklist.close()
-try:
-    output = subprocess.check_output([skslc, worklist.name], stderr=subprocess.STDOUT)
-except subprocess.CalledProcessError as err:
-    if err.returncode != 1:
-        print("### skslc error:\n")
-        print("\n".join(err.output.splitlines()))
-        sys.exit(err.returncode)
-    pass  # Compile errors (exit code 1) are expected and normal in test code
+    # Compile items one at a time.
+    if not batchCompile:
+        executeWorklist(input, worklist)
+        worklist = tempfile.NamedTemporaryFile(suffix='.worklist', delete=False)
 
-os.remove(worklist.name)
+# Compile everything all in one go.
+if batchCompile:
+    executeWorklist("", worklist)
+else:
+    worklist.close()
+    os.remove(worklist.name)
 
 # A special case cleanup pass, just for CPP and H files: if either one of these files starts with
 # `### Compilation failed`, its sibling should be replaced by an empty file. This improves clarity
